requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.3
django-crispy-forms==1.11.2
django-filter==2.4.0
django-sendgrid-v5==1.1.0
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.2.0
psycopg2==2.9.1
python-decouple==3.4
python-dotenv==0.17.1
python-http-client==3.3.2
pytz==2021.1
sendgrid==6.7.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
whitenoise==5.3.0
django-heroku == 0.3.1
djangorestframework